backend/requirements.txt
Flask==2.3.3
opencv-python-headless
requests
python-dotenv
